requirements.txt
amqp==5.3.1
asgiref==3.10.0
async-timeout==5.0.1
billiard==4.2.2
celery==5.5.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.2.7
flower==2.0.1
humanize==4.15.0
kombu==5.5.4
packaging==25.0
pillow==12.0.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2023.3.post1
redis==7.3.0
six==1.17.0
sqlparse==0.5.3
tornado==6.5.5
typing_extensions==4.8.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.14
